:broom: Tidied timeline code somewhat
